EOS-15809 Changes for CSM Setup Configure Command.
